Folded the data functions back into blast_tools. One day when we move away from blast we will no doubt make these independant of hteir tool, which will be as simple as passing in the headings for the dataframe as arguments, but they can live there for now or simplicity as they are only technically used for blast related processing.
